explanation: undo command
One of the advantages of decoupling the actions from the invoker is that we can implement easily a method to undo the action. The main idea of this method would be to perform the opposite operation the execute method does. Do notice, that this new feature is transparent for the receiver, we only need to make the changes in the invoker to be able to hold this operation.