explanation: command
The transaction interface will define what the concrete commands need to implement in order to be considered commands.

The power of the command pattern can be shown in the TransferCommand class. If we didn't implement the command patter, this responsibility should be inside the Account class, needing to receive another Account object to be able to perform the operation.