feat: initial code
We start from a code that has been already implemented, where all the logic and state transitions in our GumballMachine has been implemented inside this same class.

The problem with this code is that if we want to add a new feature we will need to add a new state and modify lots of if statements, increasing the risk of breaking the code and making add code more difficult.