explanation: add new behaviour through inheritance (wrong)
We now need to add a new behavior, flying ducks. The first approach is to create a fly method in the superclass and let all concrete ducks inherit this behavior.

The problem with this approach is that even those ducks that shouldn't fly (i.e. RubberDuck) would fly, so this is not a good solution. A localized update in the code caused a non side effect.