explanation: manage new behavior overriding (wrong)
Another approach we could follow is to leave the Duck as is, with a common fly method, and override this behavior for the ducks that can't fly.

The problem doing this is the scalability, if we add a new type of duck, i.e. DecoyDuck, that doesn't quack or fly, we need to override the methods to do nothing.

This approach leads to code duplication, difficult scalability and potentially a change in one duck could affect other duck. In conclusion, this isn't a good solution either.