explanation: manage new behavior through interfaces (wrong)
Another approach we could think of is to extract the fly and quack behavior to separate interfaces. This way we could decide which ducks needs to fly or quack.

This main problem with this idea is that we will be duplicating lots of code. If most ducks have a common fly and quack behavior, if we make a change to this behavior we should be checking all the classes. Again, this is not a good solution if we expect new types of ducks with new behaviors.