explanation: encapsulate fly and quack behavior in different classes
By separating the behavior from the Duck class we can implement all the new behaviors we want for flying and quacking without changing any exiting code.

We define two interfaces for both behaviors so we program to a supertype, allowing us to work with abstractions rather than with concrete implementations.