explanation: introduction to factory method pattern
Now we want to have more flexibility in our PizzaStore, and we want to be able to create different pizzas style depending on the region.

What we'll do is to put the create_pizza method inside the PizzaStore class again but set both the method and the class as abstract. The objective here is create concrete PizzaStores depending on the region, and letting those concrete factories how they create the pizza