explanation: implementing factory method in concrete classes
What varies among the regional PizzaStore is the style of the pizza we make. So we are moving the responsibility of deciding what pizzas to create to the concrete PizzaStores.

The key concept here is that we use the order pizza method defined in the abstract class. This method calls the create pizza, but it doesn't know what is doing behind scenes, it doesn't care what or how the pizzas are created, only that it gets a pizza. The factory method in the subclasses is the responsible of managing object creation, encapsulating its behavior and decoupling this knowledge from the client code